+++ app/models/kuhsaft/brick.rb
@@ -23,10 +23,20 @@ module Kuhsaft
               :brick_list_type,
               :presence => true
 
+    after_initialize do
+      self.position ||= has_siblings? ? brick_list.bricks.maximum(:position).to_i + 1 : 1
+    end
+
     def to_edit_partial_path
       path = self.to_partial_path.split '/'
       path << 'edit'
       path.join '/'
+    end
+
+    def has_siblings?
+      if brick_list
+        brick_list.bricks.any?
+      end
     end
 
     #